Board and list changes

+++ src/main/java/services/BoardService.java
@@ -2,6 +2,7 @@ package services;
 
 import java.util.List;
 
+
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
@@ -21,34 +22,66 @@ public class BoardService {
 	//Users can create a new board with a unique name.
 	public Board createBoard(String name) {
 		//check if the user is logged in
-//		if (User.getCurrentUser() == null) {
-//			throw new IllegalArgumentException("User not logged in");
-//		}
-		board = new Board();
-		board.setBoardName(name);
-		entityManager.persist(board);
-		return board;
+		if (User.isCurrentUser() == false) {
+			throw new IllegalArgumentException("User not logged in");
+		}
+		else {
+			// create a new board
+			board = new Board();
+			board.setBoardName(name);
+			entityManager.persist(board);
+			//User.getCurrentUser().getUserBoards().add(board);
+			return board;
+		}
 	}
 	
+
 	//Users can delete a board.
 	public void deleteBoard(long id) {
+		if (User.isCurrentUser() == false) {
+			throw new IllegalArgumentException("User not logged in");
+		}
+		else {
 		board = entityManager.find(Board.class, id);
 		entityManager.remove(board);
+		}
 	}
 	
 	//Users can view all boards they have access to.
 	public List<Board> getBoards(long id) {
-		TypedQuery<User> userBoardList = entityManager.createQuery("SELECT u from User u LEFT JOIN FETCH u.userBoards WHERE u.id = :id", User.class);
-		userBoardList.setParameter("id", id);
-		return userBoardList.getResultList().get(0).getUserBoards();
+		if (User.isCurrentUser() == false) {
+			throw new IllegalArgumentException("User not logged in");
+		}
+		else {
+		User user = entityManager.find(User.class, id);
+		return user.getUserBoards();
+		}
 	}
 	
 	//Users can invite other users to collaborate on a board.
 	public void inviteUser(long boardId, long userId) {
+		if (User.isCurrentUser() == false) {
+			throw new IllegalArgumentException("User not logged in");
+		}
+		else {
 		board = entityManager.find(Board.class, boardId);
 		User user = entityManager.find(User.class, userId);
 		user.getUserBoards().add(board);
+		board.getUsers().add(user);
+		entityManager.merge(board);
 		entityManager.merge(user);
+		}
+	}
+	
+	//Users can view all boards
+	public List<Board> getAllBoards() {
+		if (User.isCurrentUser() == false) {
+			throw new IllegalArgumentException("User not logged in");
+		}
+		else {
+		TypedQuery<Board> query = entityManager.createQuery("SELECT b from Board b LEFT JOIN FETCH b.users ", Board.class);
+		return query.getResultList();
+		}
 	}
